Change to TagLib master branch

+++ taglib/taglib-src/tests/test_bytevectorlist.cpp
@@ -23,8 +23,8 @@
  *   http://www.mozilla.org/MPL/                                           *
  ***************************************************************************/
 
-#include <tbytevector.h>
-#include <tbytevectorlist.h>
+#include "tbytevector.h"
+#include "tbytevectorlist.h"
 #include <cppunit/extensions/HelperMacros.h>
 
 using namespace std;
@@ -44,7 +44,7 @@ public:
     ByteVector v("a b");
 
     ByteVectorList l = ByteVectorList::split(v, " ");
-    CPPUNIT_ASSERT_EQUAL((size_t)2, l.size());
+    CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(2), l.size());
     CPPUNIT_ASSERT_EQUAL(ByteVector("a"), l[0]);
     CPPUNIT_ASSERT_EQUAL(ByteVector("b"), l[1]);
   }
@@ -54,7 +54,7 @@ public:
     ByteVector v("a");
 
     ByteVectorList l = ByteVectorList::split(v, " ");
-    CPPUNIT_ASSERT_EQUAL((size_t)1, l.size());
+    CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(1), l.size());
     CPPUNIT_ASSERT_EQUAL(ByteVector("a"), l[0]);
   }
